# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mashalliance API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/chat/{chat}": {
            "get": {
                "tags": [
                    "Get Chat"
                ],
                "summary": "Get Chat",
                "description": "Get Chat",
                "operationId": "getChat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "owner_id",
                        "in": "query",
                        "description": "owner_id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user_id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "owner_message_count",
                        "in": "query",
                        "description": "owner_message_count",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_message_count",
                        "in": "query",
                        "description": "user_message_count",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chat/send-message": {
            "post": {
                "tags": [
                    "Send Message"
                ],
                "summary": "Send Message",
                "description": "Send Message",
                "operationId": "sendMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message",
                                    "user_id"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "message"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "user id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/message/delete/{message}": {
            "delete": {
                "tags": [
                    "Delete Message"
                ],
                "summary": "Delete Message",
                "description": "Delete Message",
                "operationId": "deleteMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Message",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chat/delete/{chat}": {
            "delete": {
                "tags": [
                    "Delete Chat Message"
                ],
                "summary": "Delete Chat Message",
                "description": "Delete Chat Message",
                "operationId": "deleteChat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Chat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}